send_game needs chat_id by i64 not u32

+++ crates/teloxide-core/src/local_macros.rs
@@ -1301,9 +1301,9 @@ macro_rules! requester_forward {
     (@method send_game $body:ident $ty:ident) => {
         type SendGame = $ty![SendGame];
 
-        fn send_game<G>(&self, chat_id: u32, game_short_name: G) -> Self::SendGame where G: Into<String> {
+        fn send_game<G>(&self, chat_id: i64, game_short_name: G) -> Self::SendGame where G: Into<String> {
             let this = self;
-            $body!(send_game this (chat_id: u32, game_short_name: G))
+            $body!(send_game this (chat_id: i64, game_short_name: G))
         }
     };
     (@method set_game_score $body:ident $ty:ident) => {
